Move physics list wrappers into dedicated module
